requirements.txt
opencv-python~=4.10.0
scipy
pillow
librosa
OpenEXR
Pillow
numpy